+++ core/src/main/kotlin/org/jetbrains/research/testspark/core/test/TestSuiteParser.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.research.testspark.core.test.parsers
+package org.jetbrains.research.testspark.core.test
 
 import org.jetbrains.research.testspark.core.test.data.TestCaseGeneratedByLLM
 import org.jetbrains.research.testspark.core.test.data.TestSuiteGeneratedByLLM
