+++ core/src/test/kotlin/org/jetbrains/research/testspark/core/test/parsers/kotlin/KotlinJUnitTestSuiteParserTest.kt
@@ -2,6 +2,7 @@ package org.jetbrains.research.testspark.core.test.parsers.kotlin
 
 import org.jetbrains.research.testspark.core.data.JUnitVersion
 import org.jetbrains.research.testspark.core.test.data.TestSuiteGeneratedByLLM
+import org.jetbrains.research.testspark.core.test.kotlin.KotlinJUnitTestSuiteParser
 import org.jetbrains.research.testspark.core.utils.kotlinImportPattern
 import org.junit.jupiter.api.Test
 import kotlin.test.assertNotNull
